Testing to see if a post and a get can go to the same route
What:  trying to sort out if a POST and a GET go to the same place in my node server
Why:  I have a textarea that acts like a form.  I DONT want to redirect to a new page when I press enter. ( enter is a submit). This is because I have no database right now and am not storing what is on the page, so if the page redirects the person using it will lose their data.