Editing the server so it can receive data sent
What:  created an if statement that receives the data sent in a post request. Post requests send data in small packages. It's like somebody sending you one piece of a puzzle each day. The if statement receives all of these pieces of the puzzle and print some out to the consul.
Why: centrally I am going to need to store tweets on the server. This is the first step towards doing so.